update peak finder & integration

+++ nsxlib/data/DataSet.h
@@ -120,7 +120,7 @@ public:
     Eigen::MatrixXd transformedFrame(std::size_t idx);
 
     //! Return a convolved frame
-    Eigen::MatrixXi convolvedFrame(std::size_t idx, const std::string& convolver_type, const std::map<std::string,double>& parameters);
+    Eigen::MatrixXd convolvedFrame(std::size_t idx, const std::string& convolver_type, const std::map<std::string,double>& parameters);
 
     //! Get the file handle.
     void open();
